pyproject.toml
[build-system]
requires = ["uv-build>=0.1.0"]
build-backend = "uv_build"

[project]
name = "yarasp"
version = "0.1.0"
description = "Yandex Schedule API Client based on httpx and hishel, with async support, API key usage counter and limit warnings"
requires-python = ">=3.9"
authors = [
  {name = "Pavel P. Serikov", email = "linuxrd@yandex.com"}
]
maintainers = [
  {name = "Pavel P. Serikov", email = "linuxrd@yandex.com"}
]
license = {text = "MIT license"}
dependencies = [
  "typer",
  "hishel>=0.1.1,<0.2.0"
]

[project.optional-dependencies]
dev = [
    "coverage",  # testing
    "mypy",  # linting
    "pytest",  # testing
    "pytest-httpx>=0.35.0",  # testing
    "ruff"  # linting
]

[project.urls]
bugs = "https://github.com/pavelsr/yarasp/issues"
changelog = "https://github.com/pavelsr/yarasp/blob/master/changelog.md"
homepage = "https://github.com/pavelsr/yarasp"




# Mypy
# ----

[tool.mypy]
files = "."

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
allow_untyped_defs = true
disable_error_code = "attr-defined"

[dependency-groups]
dev = [
    "coverage",
    "mypy",
    "pytest",
    "pytest-httpx>=0.35.0",
    "ruff"
]

